feature/update: remove Rest API
Remove rest API update/rename rest_api example to use nfs API directly.
