+++ Cargo.toml
@@ -15,7 +15,7 @@ log = "~0.3.6"
 lru_time_cache = "~0.3.0"
 maidsafe_utilities = "~0.5.3"
 rand = "~0.3.14"
-routing = "~0.16.2"
+routing = "~0.17.0"
 rustc-serialize = "~0.3.19"
 safe_network_common = "~0.1.1"
 self_encryption = "~0.3.1"
@@ -45,7 +45,7 @@ name = "safe_client"
 
 [[example]]
 bench = false
-name = "rest_api"
+name = "nfs_api"
 
 [[example]]
 bench = false
