+++ src/nfs/mod.rs
@@ -17,8 +17,6 @@
 
 /// Module for File struct
 pub mod file;
-/// Module for Restful interfaces for storage
-pub mod rest;
 /// Errors
 pub mod errors;
 /// Helper for directory_listing and File for NFS Low level API
